requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
numpy>=2.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
plotly>=5.20.0
seaborn>=0.13.2
opencv-python>=4.10.0
transformers>=4.40.0
torch>=2.3.0
torchvision>=0.18.0
nltk>=3.9.0
speechrecognition>=3.12.0
librosa>=0.10.2
sounddevice>=0.5.0
soundfile>=0.13.0
sqlalchemy>=2.0.30
joblib>=1.4.0
Pillow>=10.4.0
requests>=2.32.0
python-dotenv>=1.0.1